member globaling work done
