+++ frontapp/my-app/app/components/loginForm.tsx
@@ -1,32 +1,37 @@
+import { useQueryClient } from "@tanstack/react-query";
 import { useState } from "react";
+import api from "../utils/api";
+import { useRouter } from "next/navigation";
 
 export default function LoginForm() {
   const [member, setMember] = useState({ username: '', password: '' })
+  const queryClient = useQueryClient();
+  const router = useRouter();
 
-
-  const loginhandleSubmit = async (e: any) => {
-    e.preventDefault();
-    const response = await fetch("http://localhost:8090/api/v1/members/login", {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(member)
-    });
-    if (response.ok) {
-      alert('로그인 성공');
-    } else {
-      alert('로그인 실패');
-    }
-  }
   const handleChange = (e: any) => {
     const { name, value } = e.target;
-    // const name: any = e.target.name;
-    // const value = e.target.value;
     setMember({ ...member, [name]: value });
     console.log({ ...member, [name]: value })
   }
+
+  const login = async (e: any) => {
+    e.preventDefault()
+
+    return await api
+      .post("/members/login", {
+        username: member.username,
+        password: member.password
+      })
+      .then(res => {
+          queryClient.setQueryData(['memberKey'], res.data.data.memberDto);
+          router.replace("/");
+      }
+      )
+      .catch((error)=>{
+        console.log(error);
+      })
+  }
+
 
   return (
     <>
@@ -43,7 +48,7 @@ export default function LoginForm() {
             </div>
           </div>
           <div className="grid gap-2">
-            <button onClick={loginhandleSubmit} className="w-full py-2 px-3 bg-indigo-600 hover:bg-indigo-700 disabled:bg-indigo-300 text-white text-sm font-semibold rounded-md shadow" type="button">로그인</button>
+            <button onClick={login} className="w-full py-2 px-3 bg-indigo-600 hover:bg-indigo-700 disabled:bg-indigo-300 text-white text-sm font-semibold rounded-md shadow" type="button">로그인</button>
             <button className="w-full py-2 px-3 bg-indigo-600 hover:bg-indigo-700 disabled:bg-indigo-300 text-white text-sm font-semibold rounded-md shadow" type="button">회원가입</button>
           </div>
         </div>
